requirements.txt
aioredis==1.2.0
amqp==2.5.0
anyjson==0.3.3
asgiref==3.1.4
asn1crypto==0.24.0
aspy.yaml==1.3.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.7.2
Automat==0.7.0
billiard==3.6.0.0
celery==4.3.0
certifi==2019.6.16
cffi==1.12.3
cfgv==2.0.0
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
confusable-homoglyphs==3.2.0
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.7
daphne==2.3.0
defusedxml==0.6.0
Django==2.2.3
django-celery==3.3.0
django-celery-beat==1.5.0
django-celery-results==1.1.2
django-crispy-forms==1.7.2
django-liqpay==1.1
django-registration==3.0.1
django-rest-swagger==2.2.0
django-templated-mail==1.1.1
django-timezone-field==3.0
djangorestframework==3.9.4
djangorestframework-jsonapi==2.8.0
djangorestframework-jwt==1.11.0
djoser==1.7.0
hiredis==1.0.0
hyperlink==19.0.0
identify==1.4.5
idna==2.8
importlib-metadata==0.18
importlib-resources==1.0.2
incremental==17.5.0
inflection==0.3.1
itypes==1.1.0
Jinja2==2.10.1
kombu==4.6.3
liqpay-python==1.0
MarkupSafe==1.1.1
msgpack==0.6.1
nodeenv==1.3.3
oauthlib==3.0.2
openapi-codec==1.3.2
Pillow==6.1.0
pre-commit==1.17.0
pycparser==2.19
PyHamcrest==1.9.0
PyJWT==1.7.1
python-crontab==2.3.8
python-dateutil==2.8.0
python3-openid==3.1.0
pytz==2019.1
PyYAML==5.1.1
redis==3.3.4
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.12.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sqlparse==0.3.0
toml==0.10.0
Twisted==19.2.1
txaio==18.8.1
uritemplate==3.0.0
urllib3==1.25.3
vine==1.3.0
virtualenv==16.6.1
zipp==0.5.2
zope.interface==4.6.0
